✨ Expose simple heureDebug/heureFin attributes
In a previous patch, we merged several "restrictions" into a unified
sensor per logical usage. It led to expose time restriction per
restriction, in attributes.
That made exploitation of time-based restriction harder to exploit
because users had to parse complex keys like "Arrosage des jardins
potagers (heureDebut)".

We simplify this for the majority of case where a single restriction is
used for a specific sensor.

This change won't chnage anything for the complex case but is breaking
for the (hopefully most common) case: we loose the "xxx (heureDebut)"
attribute in favor of "heureDebut".

Change-Id: If177ce5d8ab34ec82ba3fb208c6b5f744cfd0506