✨ Feed usage list from dataset directly
Instead of scraping vigieau website (costly, long, restricted to
mainland France cities), we now directly download the vigieau dataset.

It is super fast (less than a few seconds) and exhaustive. This is one
step to support overseas France regions.

It has also the nice side-effect of having the complete list of usages