✨ Follow HA location change
When HA instance location changes, the next vigieau refresh will detect
it and adapt.

Technically this is made by introducing a storage with all data that
used to be in config (which still are in config but it seems hard to
update this outside of the HA startup phase)

Downside: it does not yet update the entity names or the device name
which makes it a bit weird.

Fix #48